Updated Download Method
The URL which is provided by the JSON file requires a POST request to receive the URL to the actual file. I've quickly updated the download method to include the required POST request. This will be the first proper release, not marked as a pre-release. Please report any issues you encounter! Thank you for using Memoire!