Rebuilt as Release
Rebuilt the application in the Release channel in Visual Studio. Now you're able to find the application in both the Debug and Release channels, making it easier to modify or work with the code.

Also included the .bat files for launching the two alternate versions of the application in the Debug channel and their compiled executables in the Release channel.